Touching & Swiping?
Katrin's little thing isn't going so well. We got the half / half thing to work well, but then realized that it wouldn't scroll. Damn! So we implemented touchswipe.js, which worked wonderfully -- but it's not very performant on an iPhone. Maybe if we scroll the master div instead? We'll see. So close, and yet so far!